Refactor imports for consistency: update import paths in game and content_manager modules, remove redundant binds init file
