+++ CityOfBinds/__init__.py
@@ -1,6 +1,14 @@
 # /CityOfBinds/__init__.py
-from .src.game.binds import Bind, WASDBind, iWASDBind
-from .src.game import BindFile, RotatingBind, RandomBinds, RandomWalk, WASDRotatingBind
+from .src.game import (
+    Bind,
+    WASDBind,
+    iWASDBind,
+    BindFile,
+    RotatingBind,
+    RandomBinds,
+    RandomWalk,
+    WASDRotatingBind,
+)
 from .src.content_managers import (
     BindTemplate,
     WASDBindTemplate,
