+++ CityOfBinds/src/game/__init__.py
@@ -1,10 +1,16 @@
+from .binds.bind import Bind
+from .binds.wasd_binds import WASDBind, iWASDBind
 from .bind_file.bind_file import BindFile
 from .bind_file.constants import BindFileConstants
 from .rotating_binds.rotating_bind import RotatingBind
 from .rotating_binds.wasd_rotating_bind import WASDRotatingBind
 from .rotating_binds.random_binds import RandomBinds, RandomWalk
 
+
 __all__ = [
+    "Bind",
+    "WASDBind",
+    "iWASDBind",
     "BindFile",
     "BindFileConstants",
     "RotatingBind",
