--- CityOfBinds/src/game/binds/__init__.py
@@ -1,4 +1,0 @@
-from .bind import Bind
-from .wasd_binds import WASDBind, iWASDBind
-
-__all__ = ["Bind", "WASDBind", "iWASDBind"]
